Move MyApp to app.dart
