+++ lib/main.dart
@@ -1,8 +1,6 @@
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'package:pasta_watford/app/features/home/home_page.dart';
-import 'package:pasta_watford/app/features/login/login_page.dart';
+import 'package:pasta_watford/app/app.dart';
 import 'firebase_options.dart';
 
 void main() async {
@@ -11,42 +9,4 @@ void main() async {
     options: DefaultFirebaseOptions.currentPlatform,
   );
   runApp(const MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        colorScheme: ColorScheme.fromSeed(
-          seedColor: const Color.fromARGB(255, 54, 244, 89),
-        ),
-        useMaterial3: false,
-      ),
-      home: const RootPage(),
-    );
-  }
-}
-
-class RootPage extends StatelessWidget {
-  const RootPage({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return StreamBuilder<User?>(
-      stream: FirebaseAuth.instance.authStateChanges(),
-      builder: (context, snapshot) {
-        final user = snapshot.data;
-        if (user == null) {
-          return LoginPage();
-        }
-        return HomePage(user: user);
-      },
-    );
-  }
 }
